requirements.txt
aiogram
pydantic-settings
pydantic
tortoise-orm[asyncpg]
kurigram
tgcrypto
python-dotenv